--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/field/TeamElementLocation.java
@@ -1,23 +1,0 @@
-package org.firstinspires.ftc.teamcode.field;
-
-/**
- * Location of the team element on the field.
- */
-public enum TeamElementLocation {
-    /**
-     * The location of the team element is not known
-     */
-    UNKNOWN,
-    /**
-     * The team element is on the left spike mark
-     */
-    LEFT_SPIKE_MARK,
-    /**
-     * The team element is on the middle spike mark
-     */
-    MIDDLE_SPIKE_MARK,
-    /**
-     * The team element is on the right spike mark
-     */
-    RIGHT_SPIKE_MARK
-}
